Restore linux 32bit cross build smoker runs.
The smoker failures in 32bit compiles on AMD64 systems were due to
an out of date package cache in the ubuntu images rather than problems
with the Perl cross-building support.

This restores the cross building tests as a separate workflow that
mirrors the matrix of configurations tested with native AMD64 compiles.